# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bengkel Sampah API",
        "description": "Dokumentasi API untuk sistem OTP dan autentikasi Bengkel Sampah",
        "contact": {
            "name": "Tim Bengkel Sampah",
            "email": "support@bengkelsampah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Dapatkan daftar event aktif",
                "description": "Mengembalikan daftar event aktif dengan paginasi, diurutkan dari yang terbaru dibuat",
                "operationId": "getEvents",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nomor halaman untuk paginasi",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Bersih-bersih Pantai"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Kegiatan membersihkan pantai dari sampah plastik..."
                                                            },
                                                            "cover": {
                                                                "type": "string",
                                                                "example": "https://example.com/cover.jpg"
                                                            },
                                                            "start_datetime": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "end_datetime": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "location": {
                                                                "type": "string",
                                                                "example": "Pantai Indah"
                                                            },
                                                            "max_participants": {
                                                                "type": "integer",
                                                                "example": 100
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "active"
                                                            },
                                                            "participants_count": {
                                                                "type": "integer",
                                                                "example": 25
                                                            },
                                                            "has_result": {
                                                                "type": "boolean",
                                                                "example": false
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            },
                                                            "admin_name": {
                                                                "type": "string",
                                                                "example": "Admin Bengkel Sampah"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "has_more_pages": {
                                                            "type": "boolean",
                                                            "example": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak terautentikasi. Token tidak diberikan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/events/{id}/toggle-join": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Join atau unjoin event",
                "description": "Toggle status join user ke event. Jika sudah join akan unjoin, jika belum join akan join",
                "operationId": "toggleEventJoin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil join event"
                                        },
                                        "data": {
                                            "properties": {
                                                "action": {
                                                    "description": "joined atau unjoined",
                                                    "type": "string",
                                                    "example": "joined"
                                                },
                                                "user_has_joined": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Event tidak aktif atau sudah penuh",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event tidak aktif atau sudah penuh"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak terautentikasi. Token tidak diberikan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}